seprated main- Question_Managment now is a blueprint- modified tamplates to connect properly to blueprint endpoints
