+++ templates/add_category.html
@@ -2,7 +2,7 @@
 
 {% block content %}
 
-<form action="{{ url_for('add_category') }}" method="post">
+<form action="{{ url_for('Question_Management.add_category') }}" method="post">
     <div class="mb-3">
         <div class="col-md-4">
             <label for="new_category" class="form-label">New Category</label>
@@ -12,7 +12,7 @@
     <button type="submit" class="btn btn-primary">Submit</button>
 
 </form>
-<a href="{{ url_for('show_categories') }}" class="btn">Return</a>
+<a href="{{ url_for('Question_Management.show_categories') }}" class="btn">Return</a>
 <br>
 {% include 'flash_.html' %}
