+++ templates/categories.html
@@ -5,7 +5,7 @@
   <ul class="list-group">
     {% for category in categories %}
       <li class="list-group-item">
-        <a href="{{ url_for('take_quiz', category=category) }}">{{ category }}</a>
+        <a href="{{ url_for('Question_Management.take_quiz', category=category) }}">{{ category }}</a>
       </li>
     {% endfor %}
   </ul>
