+++ templates/layout.html
@@ -19,9 +19,9 @@
           <div class="navbar-nav">
             <a class="nav-link" href="{{ url_for('index') }}">Home</a>
             <a class="nav-link active" aria-current="page" href="#">User/LogIn</a>
-            <a class="nav-link" href="{{ url_for('show_categories') }}">Questions</a>
+            <a class="nav-link" href="{{ url_for('Question_Management.show_categories') }}">Questions</a>
             <a class="nav-link" href="#">Quiz</a>
-            <a class="nav-link" href="{{ url_for('profile') }}">Profile</a>
+            <a class="nav-link" href="{{ url_for('Question_Management.profile') }}">Profile</a>
             <a class="nav-link disabled" aria-disabled="true">Disabled</a>
           </div>
         </div>
