+++ templates/show_categories.html
@@ -14,7 +14,7 @@
     <div class="col-md-1 mb-1">
         <div class="card">
             <div class="card-body text-center">
-                <a href="{{ url_for('Question_manager', category=category.name) }}" class="category-link">{{ category.name }}</a>
+                <a href="{{ url_for('Question_Management.Question_manager', category=category.name) }}" class="category-link">{{ category.name }}</a>
             </div>
         </div>
     </div>
@@ -22,7 +22,7 @@
     <div class="col-md-1 mb-1">
         <div class="card add-category">
             <div class="card-body text-center">
-                <a href="{{ url_for('add_category') }}" class="category-link">+</a>
+                <a href="{{ url_for('Question_Management.add_category') }}" class="category-link">+</a>
             </div>
         </div>
     </div>
